# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Survey Service OpenApi3 Documentation",
        "description": "Behandam Survey Service Api documentation",
        "contact": {
            "email": "admin@kermany.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "LocalHost API Server"
        },
        {
            "url": "https://debug.behaminplus.ir/survey-service",
            "description": "Debug API Server"
        },
        {
            "url": "https://develop.behaminplus.ir/survey-service",
            "description": "DEVELOP API Server"
        }
    ],
    "paths": {
...
        "/api/v1/campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "get list of all campaigns",
                "operationId": "getListOfCampaigns",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "define a new campaign",
                "operationId": "postANewCampaign",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CampaignStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/campaigns/{campaignId}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "get item by id",
                "operationId": "getCampaignItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "campaignId",
                        "in": "path",
                        "description": "ID of campaign",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "update campaign item by id",
                "operationId": "updateCampaignItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "campaignId",
                        "in": "path",
                        "description": "ID of campaign",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CampaignStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "remove item by id",
                "operationId": "removeAnItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "campaignId",
                        "in": "path",
                        "description": "ID of campaign",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/api/v1/answers/{answerId}": {
            "delete": {
                "tags": [
                    "Answers"
                ],
                "summary": "remove and app by id",
                "operationId": "removeAnItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "ID of answer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/api/v1/packages/{packageId}": {
            "get": {
                "tags": [
                    "Packages"
                ],
                "summary": "get package item by id",
                "operationId": "getPackageItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "ID of package",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Packages"
                ],
                "summary": "update package item by id",
                "operationId": "updatepackageItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "ID of package",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PackageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Packages"
                ],
                "summary": "remove and app by id",
                "operationId": "removeAnItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "ID of package",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/api/v1/choices/{choiceId}/questions": {
            "get": {
                "tags": [
                    "Choices"
                ],
                "summary": "get choice item by choice id",
                "operationId": "getChoiceItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "choiceId",
                        "in": "path",
                        "description": "ID of choice",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Choices"
                ],
                "summary": "update choice item by id",
                "operationId": "updatechoiceItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "choiceId",
                        "in": "path",
                        "description": "ID of choice",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChoiceStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Choices"
                ],
                "summary": "remove and app by id",
                "operationId": "removeAnItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "choiceId",
                        "in": "path",
                        "description": "ID of choice",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/api/v1/questions/{questionId}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "get question item by id",
                "operationId": "getItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "ID of question",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "update Question item by id",
                "operationId": "updateQuestionItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "ID of Question",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "remove and app by id",
                "operationId": "removeAnItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "ID of question",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CampaignStoreRequest": {
                "title": "CampaignStoreRequest",
                "description": "Campaign Store Request body data",
                "required": [
                    "title"
                ],
                "type": "object",
                "example": {
                    "title": "YOUR TITLE OF ..."
                }
            },
...
            "ChoiceStoreRequest": {
                "title": "ChoiceStoreRequest",
                "description": "Choices Store Request body data",
                "required": [
                    "title"
                ],
                "type": "object",
                "example": {
                    "title": "YOUR TITLE OF ...2",
                    "weight": 2,
                    "is_active": 1,
                    "order": 1
                }
            },
...
            "PackageRequest": {
                "title": "PackageRequest",
                "description": "Package Request",
                "required": [
                    "title"
                ],
                "type": "object",
                "example": {
                    "title": "نام پکیج",
                    "package_type_id": 1,
                    "client_app_id": 1,
                    "parent_id": null,
                    "description": "description",
                    "first_text": "first_text",
                    "final_text": "final_text",
                    "started_at": "2021-06-02T14:48:37",
                    "finished_at": "2021-06-02T14:48:37",
                    "is_active": 1,
                    "is_deletable": 1
                }
            },
            "QuestionStoreRequest": {
                "title": "QuestionStoreRequest",
                "description": "Question Store Request body data",
                "required": [
                    "title",
                    "package_id"
                ],
                "type": "object",
                "example": {
                    "title": "YOUR TITLE OF ...",
                    "package_id": 1,
                    "weight": 2,
                    "is_active": 1,
                    "order": 1
                }
            },
...
        },
        "securitySchemes": {
            "access-token": {
                "type": "http",
                "description": "Use Custom access token",
                "name": "AccessToken",
                "scheme": "access-token"
            }
        }
    },
...
    "security": [
        []
    ]
}